rename BpmnOverlay file for clarity
