+++ src/component/mxgraph/overlay/custom-overlay.ts
