+++ test/integration/matchers/matcher-utils.ts
@@ -18,7 +18,7 @@ import MatcherContext = jest.MatcherContext;
 import CustomMatcherResult = jest.CustomMatcherResult;
 import { mxgraph } from '../../../src/component/mxgraph/initializer';
 import { mxCell, mxGeometry, StyleMap } from 'mxgraph';
-import { MxGraphCustomOverlay } from '../../../src/component/mxgraph/overlay/BpmnOverlay';
+import { MxGraphCustomOverlay } from '../../../src/component/mxgraph/overlay/custom-overlay';
 
 export interface ExpectedStateStyle extends StyleMap {
   verticalAlign?: string;
